% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{calculateMagicRainbowStars}
\alias{calculateMagicRainbowStars}
\title{calculateMagicRainbowStars}
\usage{
calculateMagicRainbowStars(p_value)
}
\arguments{
\item{p_value}{Obviously the p_value.}
}
\value{
Returns: Signif. codes:  0 "***" 0.001 "**" 0.01 "*" 0.05 "." 0.1 " " 1
}
\description{
Returns different notations depending which value the p_value holds.
}
