% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{cPrint}
\alias{cPrint}
\title{cPrint (custom print)}
\usage{
cPrint(x, stripoff = FALSE)
}
\arguments{
\item{x}{Any object that normally can be printed.}

\item{stripoff}{If set to TRUE and \code{x} is a data.frame, the column names will be stripped off.}
}
\value{
Nothing.
}
\description{
Prints the value in \code{x}. This class can be used to print inside RC classes, which is not possible otherwise.
}
