% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg Class}
\value{
Nothing.
}
\description{
This class does linear regression and provided different methods to display the data.
}
\section{Fields}{

\describe{
\item{\code{l_X}}{matrix.}

\item{\code{l_y}}{matrix.}

\item{\code{l_beta}}{matrix.}

\item{\code{l_y_fitted_values}}{matrix.}

\item{\code{l_e}}{matrix.}

\item{\code{l_n}}{numeric.}

\item{\code{l_p}}{numeric.}

\item{\code{l_df}}{numeric.}

\item{\code{l_sigma_s}}{matrix.}

\item{\code{l_var_beta}}{matrix.}

\item{\code{l_formula}}{formula.}

\item{\code{l_p_values}}{numeric.}

\item{\code{l_data_set_name}}{character.}
}}


