% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{liu_theme}
\alias{liu_theme}
\title{Liu Theme}
\usage{
liu_theme(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{The base size of the font.}

\item{base_family}{The base family of the font.}
}
\value{
The Liu Theme. Add this to our plot.
}
\description{
Liu Theme
}
