% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg Class}
\value{
Nothing.
}
\description{
This class does linear regression and provided different methods to display the data.
}
\section{Fields}{

\describe{
\item{\code{l_X}}{matrix. Independent Values.}

\item{\code{l_y}}{matrix. Depenent Values.}

\item{\code{l_beta}}{matrix. Regression Coeffcients.}

\item{\code{l_y_fitted_values}}{matrix. Fitted Values.}

\item{\code{l_e}}{matrix. Residuals.}

\item{\code{l_n}}{numeric. Number of Independant Values (Number of Observations).}

\item{\code{l_p}}{numeric. Number of Dependant Values (Number of Parameters in the Model).}

\item{\code{l_df}}{numeric. Degrees of freedom.}

\item{\code{l_sigma_s}}{matrix. Residual Variance.}

\item{\code{l_var_beta}}{matrix. Variance of the Regression Coefficients.}

\item{\code{l_formula}}{formula. The Formula for the Linear Regression.}

\item{\code{l_p_values}}{numeric. P-Values.}

\item{\code{l_t_beta}}{matrix. T-Values for each Coefficient.}

\item{\code{l_data_set_name}}{character. The Given Data.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the Coefficients}

\item{\code{initialize(formula, data)}}{Constructor for creating the object. Arguments are the formlua and the corresponding data frame.}

\item{\code{plot()}}{Plots the Fitted Values one time with the Residuals and one time with standardized Residuals}

\item{\code{pred()}}{Returns the Fitted Values}

\item{\code{print()}}{Prints some basic information like the coefficients.}

\item{\code{resid()}}{Returns the Residuals}

\item{\code{summary()}}{Prints a summary of the Linear Regression}
}}

